Fixed issue with BackgroundChannelRates
This commit fixes an error that arises when a BackgroundChannelRates object first has its channels integrated into one and is then time sliced.

Change consists of:

 Adjusting the __init__ method dimension checks to allow for one dimensional rates and errors. This is not needed in the TimeChannelBins class due to the integrate forming a TimeBins object.

Changing the way the counts array is calculated in the 1d case to ensure the arrays stay 1d.

Adjusting the slice time method to only index one dimension if the rates are one dimensional.