Add background lightcurves and spectra for non energy calibrated data  (#79)
* Added non energy calibrated background functionality, Only currently tested for lightcurve fitting

* Updated doc strings and fixed issues with background channel spectra

* Removed comment regarding docstring

Issue mentioned in the comment has since been fixed.

* Fixed minor docstring error

* Unit Tests added and minor methods fixes

This commit adds the required unit tests for the new uncalibrated background classes. These are almost entirely a copy and paste of their calibrated counterparts with only changes made to mentions of energies as well as removal of bak methods.

Minor fixes were also made to BackgroundChannelSpectrum init method due to issues discovered while testing.

* Added Documentation

* Fixed issue with BackgroundChannelRates

This commit fixes an error that arises when a BackgroundChannelRates object first has its channels integrated into one and is then time sliced.

Change consists of:

 Adjusting the __init__ method dimension checks to allow for one dimensional rates and errors. This is not needed in the TimeChannelBins class due to the integrate forming a TimeBins object.

Changing the way the counts array is calculated in the 1d case to ensure the arrays stay 1d.

Adjusting the slice time method to only index one dimension if the rates are one dimensional.

---------

Signed-off-by: Bill Cleveland <120688282+BillCleveland-USRA@users.noreply.github.com>
Co-authored-by: Bill Cleveland <120688282+BillCleveland-USRA@users.noreply.github.com>